include/soc/am6x/resasg_types.h: Fix MISRA.ETYPE.CATEGORY.DIFFERENT.2012
Fix some MISRA.ETYPE.CATEGORY.DIFFERENT.2012 issues.

Fixes: SYSFW-1510

+++ include/soc/am6x/resasg_types.h
@@ -57,8 +57,8 @@
  */
 
 #define RESASG_UTYPE(type, subtype) \
-	(((type << RESASG_TYPE_SHIFT) & RESASG_TYPE_MASK) | \
-	 ((subtype << RESASG_SUBTYPE_SHIFT) & RESASG_SUBTYPE_MASK))
+	((((u16) (type) << RESASG_TYPE_SHIFT) & RESASG_TYPE_MASK) |	\
+	 (((u16) (subtype) << RESASG_SUBTYPE_SHIFT) & RESASG_SUBTYPE_MASK))
 
 /**
  * IA subtypes definitions
